Avoid canon-case check when we are not via_notify
If we're stat_path'ing as part of a walk then we assume that
the path that we're operating on is already the canonical name.

The idea is that we just checked and know that the name is in canonical
case, either because we just walked the dir or just examined it in
response to a notification.  In this case we can avoid an extraneous
operation.